Encode lengths as int64 instead of uint32

+++ tape/lib/src/blocks/blocks/bytes_block.dart
@@ -17,11 +17,11 @@ class BytesBlock implements Block {
 
 // An encoded [BytesBlock] looks like this:
 // | num bytes | byte | byte | byte | ... |
-// The number of bytes is encoded as a uint32, the bytes as uint8s.
+// The number of bytes is encoded as an int64, the bytes as uint8s.
 
 extension _BytesBlockWriter on _Writer {
   void writeBytesBlock(BytesBlock block) {
-    writeUint32(block.bytes.length);
+    writeInt64(block.bytes.length);
     block.bytes.forEach(writeUint8);
   }
 }
@@ -29,7 +29,7 @@ extension _BytesBlockWriter on _Writer {
 extension _BytesBlockReader on _Reader {
   // TODO: We can probably make this lots more efficient by just returning a view of the actual ByteData. To be able to do that, we'd have the change the Reader signature though.
   BytesBlock readBytesBlock() {
-    final length = readUint32();
+    final length = readInt64();
     return BytesBlock([
       for (var i = 0; i < length; i++) readUint8(),
     ]);
