+++ tape/lib/src/blocks/blocks/list_block.dart
@@ -14,18 +14,18 @@ class ListBlock implements Block {
 
 // An encoded [ListBlock] looks like this:
 // | length | item | item | ... |
-// The length is encoded as a uint32.
+// The length is encoded as an int64.
 
 extension _ListBlockWriter on _Writer {
   void writeListBlock(ListBlock block) {
-    writeUint32(block.children.length);
+    writeInt64(block.children.length);
     block.children.forEach(writeBlock);
   }
 }
 
 extension _ListBlockReader on _Reader {
   ListBlock readListBlock() {
-    final length = readUint32();
+    final length = readInt64();
     return ListBlock([
       for (var i = 0; i < length; i++) readBlock(),
     ]);
